requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.4
django-environ==0.4.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==21.3
psycopg2-binary==2.9.3
PyJWT==1.7.1
pyparsing==3.0.9
pytz==2020.4
PyYAML==5.3.1
requests==2.25.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
